+++ ActionUtilities macOS/Extensions/ColorExtensions.swift
@@ -99,6 +99,279 @@ extension NSColor {
         self.init(red: r, green: g, blue: b, alpha: a)
     }
     
+    /**
+     Initializes a `NSColor` for the given grayscale percent and alph value.
+     
+     ## Example:
+     ```swift
+     let gray = NSColor(fromGrayScaleShade: 200)
+     ```
+     
+     - Parameters:
+         - shade: A number from 0 to 255 specifying the shade of gray to mix.
+         - alphaPercent: A number from 0 to 100 specifying the opacity of the color.
+     */
+    public convenience init(fromGrayScaleShade shade:Int, withAlphaPercent alphaPercent: Int = 100) {
+        let value = CGFloat(shade) / 255.0
+        let alphaValue = CGFloat(alphaPercent) / 100.0
+        self.init(red: value, green: value, blue: value, alpha: alphaValue)
+    }
+    
+    /**
+     Initializes a `NSColor` from the given red, green, blue and alpha values.
+     
+     ## Example:
+     ```swift
+     let white = NSColor(red: 255, green: 255, blue: 255)
+     ```
+     
+     - Parameters:
+         - red: A number between 0 and 255.
+         - green: A number between 0 and 255.
+         - blue: A number between 0 and 255.
+         - alpha: A percentage between 0 and 100.
+     */
+    public convenience init(red: Int, green: Int, blue: Int, alpha: Int = 100) {
+        let r = CGFloat(red) / 255.0
+        let g = CGFloat(green) / 255.0
+        let b = CGFloat(blue) / 255.0
+        let a = CGFloat(alpha) / 100.0
+        
+        self.init(red: r, green: g, blue: b, alpha: a)
+    }
+    
+    /**
+     Initialize a `NSColor` from the given tuple of red, green, blue and alpha values where:
+     
+     * `red` - Is a number between 0 and 255.
+     * `green` - Is a number between 0 and 255.
+     * `blue` - Is a number between 0 and 255.
+     * `alpha` - Is a percentage between 0 and 100.
+     
+     ## Example:
+     ```swift
+     let white = NSColor(fromRGBA: (red: 255, green: 255, blue: 255, alpha: 100))
+     ```
+     
+     - Parameter rgbComponents: A tuple of red, green, blue and alpha values.
+     */
+    public convenience init (fromRGBA rgbComponents: (red: Int, green: Int, blue: Int, alpha: Int)) {
+        let r = CGFloat(rgbComponents.red) / 255.0
+        let g = CGFloat(rgbComponents.green) / 255.0
+        let b = CGFloat(rgbComponents.blue) / 255.0
+        let a = CGFloat(rgbComponents.alpha) / 100.0
+        
+        self.init(red: r, green: g, blue: b, alpha: a)
+    }
+    
+    /**
+     Initializes a `NSColor` from the given hue, saturation, brightness and alpha values.
+     
+     ## Example:
+     ```swift
+     let red = NSColor(hue: 0, saturation: 100, brightness: 100)
+     ```
+     
+     - Parameters:
+         - hue: A number between 0 and 360.
+         - saturation: A percentage between 0 and 100.
+         - brightness: A percentage between 0 and 100.
+         - alpha: A percentage between 0 and 100.
+     */
+    public convenience init (hue: Int, saturation: Int, brightness: Int, alpha: Int = 100) {
+        let h = CGFloat(hue) / 360.0
+        let s = CGFloat(saturation) / 100.0
+        let b = CGFloat(brightness) / 100.0
+        let a = CGFloat(alpha) / 100.0
+        
+        self.init(hue: h, saturation: s, brightness: b, alpha: a)
+    }
+    
+    /**
+     Initializes a `NSColor` from the given tuple of hue, saturation, brightness and alpha values where:
+     
+     * `hue` - Is a number between 0 and 360.
+     * `saturation` - Is a percentage between 0 and 100.
+     * `brightness` - Is a percentage between 0 and 100.
+     * `alpha` - Is a percentage between 0 and 100.
+     
+     ## Example:
+     ```swift
+     let red = NSColor(fromHSBA: (hue: 0, saturation: 100, brightness: 100, alpha: 100))
+     ```
+     
+     - Parameter hsbComponents: A tuple of hue, saturation, brightness and alpha values.
+     */
+    public convenience init(fromHSBA hsbComponents: (hue: Int, saturation: Int, brightness: Int, alpha: Int)) {
+        let h = CGFloat(hsbComponents.hue) / 360.0
+        let s = CGFloat(hsbComponents.saturation) / 100.0
+        let b = CGFloat(hsbComponents.brightness) / 100.0
+        let a = CGFloat(hsbComponents.alpha) / 100.0
+        
+        self.init(hue: h, saturation: s, brightness: b, alpha: a)
+    }
+    
+    // MARK: - Computed Properties
+    /**
+     Returns the red, green, blue and alpha components of the given color where:
+     
+     * `red` - Is a number between 0 and 255.
+     * `green` - Is a number between 0 and 255.
+     * `blue` - Is a number between 0 and 255.
+     * `alpha` - Is a percentage between 0 and 100.
+     
+     ## Example:
+     ```swift
+     let components = NSColor.red.rgbComponents
+     print(components.red)
+     ```
+     */
+    public var rgbComponents: (red: Int, green: Int, blue: Int, alpha: Int) {
+        get {
+            var (red, green, blue, alpha) = (CGFloat(0.0), CGFloat(0.0), CGFloat(0.0), CGFloat(0.0))
+            self.getRed(&red, green: &green, blue: &blue, alpha: &alpha)
+            return (red: Int(red * 255), green: Int(green * 255), blue: Int(blue * 255), alpha: Int(alpha * 100))
+        }
+    }
+    
+    /**
+     Returns the alpha value of the color as a percentage between 0 and 100.
+     
+     ## Example:
+     ```swift
+     let alpha = NSColor.red.alphaValue
+     ```
+     */
+    public var alphaValue: Int {
+        get {return rgbComponents.alpha}
+    }
+    
+    /**
+     Returns the red value of the color as a number between 0 and 255.
+     
+     ## Example:
+     ```swift
+     let red = NSColor.red.redValue
+     ```
+     */
+    public var redValue: Int {
+        get {return rgbComponents.red}
+    }
+    
+    /**
+     Returns the green value of the color as a number between 0 and 255.
+     
+     ## Example:
+     ```swift
+     let green = NSColor.red.greenValue
+     ```
+     */
+    public var greenValue: Int {
+        get {return rgbComponents.green}
+    }
+    
+    /**
+     Returns the blue value of the color as a number between 0 and 255.
+     
+     ## Example:
+     ```swift
+     let blue = NSColor.red.blueValue
+     ```
+     */
+    public var blueValue: Int {
+        get {return rgbComponents.blue}
+    }
+    
+    /**
+     Returns the hue, saturation, brightness (value) and alpha component of the given color where:
+     
+     * `hue` - Is a number between 0 and 360.
+     * `saturation` - Is a percentage between 0 and 100.
+     * `brightness` - Is a percentage between 0 and 100.
+     * `alpha` - Is a percentage between 0 and 100.
+     
+     ## Example:
+     ```swift
+     let components = NSColor.red.hsbComponents
+     print(components.hue)
+     ```
+     */
+    public var hsbComponents: (hue: Int, saturation: Int, brightness: Int, alpha: Int) {
+        get {
+            var (hue, saturation, brightness, alpha) = (CGFloat(0.0), CGFloat(0.0), CGFloat(0.0), CGFloat(0.0))
+            self.getRed(&hue, green: &saturation, blue: &brightness, alpha: &alpha)
+            return (hue: Int(hue * 360), saturation: Int(saturation * 100), brightness: Int(brightness * 100), alpha: Int(alpha * 100))
+        }
+    }
+    
+    /**
+     Returns the hue value of the color as a number between 0 and 360.
+     
+     ## Example:
+     ```swift
+     let hue = NSColor.red.hueValue
+     ```
+     */
+    public var hueValue: Int {
+        get { return hsbComponents.hue}
+    }
+    
+    /**
+     Returns the saturation value of the color as a percentage between 0 and 100.
+     
+     ## Example:
+     ```swift
+     let saturation = NSColor.red.saturationValue
+     ```
+     */
+    public var saturationValue: Int {
+        get {return hsbComponents.saturation}
+    }
+    
+    /**
+     Returns the brightness value of the color as a percentage between 0 and 100.
+     
+     ## Example:
+     ```swift
+     let brightness = NSColor.red.brightnessValue
+     ```
+     */
+    public var brightnessValue: Int {
+        get {return hsbComponents.brightness}
+    }
+    
+    /**
+     Returns the shade and alpha of a given gray scale color where:
+     
+     * `shade` - Is a number between 0 and 255.
+     * `alpha` - Is a percentage between 0 and 100.
+     
+     ## Example:
+     ```swift
+     let components = NSColor.gray.grayScaleComponents
+     print(components.shade)
+     ```
+     */
+    public var grayScaleComponents: (shade: Int, alpha: Int) {
+        get {
+            var (red, green, blue, alpha) = (CGFloat(0.0), CGFloat(0.0), CGFloat(0.0), CGFloat(0.0))
+            self.getRed(&red, green: &green, blue: &blue, alpha: &alpha)
+            return (shade: Int(red * 255), alpha: Int(alpha * 100))
+        }
+    }
+    
+    /**
+     Returns the gray scale shade of the color as a number between 0 and 255.
+     
+     ## Example:
+     ```swift
+     let shade = NSColor.gray.shadeValue
+     ```
+     */
+    public var shadeValue: Int {
+        get { return grayScaleComponents.shade}
+    }
+    
     // MARK: - Public Functions
     /**
      Converts a `NSColor` to a hex string in the format `rrggbb` or `rrggbbaa` where:
@@ -144,5 +417,117 @@ extension NSColor {
         } else {
             return prefix + String(format: "%02lX%02lX%02lX", lroundf(r * 255), lroundf(g * 255), lroundf(b * 255))
         }
+    }
+    
+    /**
+     Returns a color based off of this color with the given alpha value.
+     
+     ## Example:
+     ```swift
+     let color = NSColor.white.withAlphaValue(50)
+     ```
+     
+     - Parameter alpha: A percentage between 0 and 100.
+     */
+    public func withAlphaValue(_ alpha: Int) -> NSColor {
+        var components = rgbComponents
+        components.alpha = alpha
+        return NSColor(fromRGBA: components)
+    }
+    
+    /**
+     Returns a color based off of this color with the given red value.
+     
+     ## Example:
+     ```swift
+     let color = NSColor.white.withRedValue(255)
+     ```
+     
+     - Parameter red: A number between 0 and 255.
+     */
+    public func withRedValue(_ red: Int) -> NSColor {
+        var components = rgbComponents
+        components.red = red
+        return NSColor(fromRGBA: components)
+    }
+    
+    /**
+     Returns a color based off of this color with the given green value.
+     
+     ## Example:
+     ```swift
+     let color = NSColor.white.withGreenValue(255)
+     ```
+     
+     - Parameter green: A number between 0 and 255.
+     */
+    public func withGreenValue(_ green: Int) -> NSColor {
+        var components = rgbComponents
+        components.green = green
+        return NSColor(fromRGBA: components)
+    }
+    
+    /**
+     Returns a color based off of this color with the given blue value.
+     
+     ## Example:
+     ```swift
+     let color = NSColor.white.withBlueValue(255)
+     ```
+     
+     - Parameter blue: A number between 0 and 255.
+     */
+    public func withBlueValue(_ blue: Int) -> NSColor {
+        var components = rgbComponents
+        components.blue = blue
+        return NSColor(fromRGBA: components)
+    }
+    
+    /**
+     Returns a color based off of this color with the given hue value.
+     
+     ## Example:
+     ```swift
+     let color = NSColor.white.withHueValue(100)
+     ```
+     
+     - Parameter hue: A number between 0 and 360.
+     */
+    public func withHueValue(_ hue: Int) -> NSColor {
+        var components = hsbComponents
+        components.hue = hue
+        return NSColor(fromHSBA: components)
+    }
+    
+    /**
+     Returns a color based off of this color with the given saturation value.
+     
+     ## Example:
+     ```swift
+     let color = NSColor.white.withSaturationValue(100)
+     ```
+     
+     - Parameter saturation: A percentage between 0 and 100.
+     */
+    public func withSaturationValue(_ saturation: Int) -> NSColor {
+        var components = hsbComponents
+        components.saturation = saturation
+        return NSColor(fromHSBA: components)
+    }
+    
+    /**
+     Returns a color based off of this color with the given brightness value.
+     
+     ## Example:
+     ```swift
+     let color = NSColor.white.withBrightnessValue(100)
+     ```
+     
+     - Parameter brightness: A percentage between 0 and 100.
+     */
+    public func withBrightnessValue(_ brightness: Int) -> NSColor {
+        var components = hsbComponents
+        components.brightness = brightness
+        return NSColor(fromHSBA: components)
     }
 }